refactor: RoundResult를 vo 패키지에서 dto 패키지로 이동
