+++ src/test/java/racingcar/dto/ResponseRoundResultDtoTest.java
@@ -6,7 +6,6 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.vo.RoundResult;
 
 public class ResponseRoundResultDtoTest {
